(UNTESTED) Restructured code layout, FKs can be null
The execution now flows through more functions, with  the the main
 'populate.php' code broken into 2 new files.
* [parser](https://github.com/LAripping/NetManager/blob/master/data/parser.php) now contains all the parser-related code and...
* [population_functions](https://github.com/LAripping/NetManager/blob/master/data/populate_functions.php) holds the internals so that...
* [populate](https://github.com/LAripping/NetManager/blob/master/data/populate.php) is left as the processing frontend (it will serve and process a form in a later commit).

FKs in device->wlan and wlan->device can now be null.
It just perplexes the process of 'insert as it comes' so much that
tricks like transactios must be used, and a careful update planning...
Now i can just fill them with NULL until their values become available.