3NF, Troubleshoot 'Populate DB',new page 'Empty DB'
The most important change is that now the device and wlan data are
_first_ processed and inserted if needed, so that packet rows inserted
later will satisfy the foreign key constraints for source-/dest-hw-addr.

Also, the multivalue-attribute's-table 'packet_has_protocol' is now a
single attribute containing a string listing the protocols found. Aside
from making the DB in a 3rd Normal Form, it greatly reduces the total
number of rows in the DB. Per-packet protocols can still be tracked by
searching in this string.