Forms for empty,populate.Works if no wlan is found
Forms have been created as frontend for the 'Empty DB' option
(to avoid accidental emptying) and the 'Populate DB' option (to select
the file to be parsed from the webroot.

The 'Populate DB' operation currently works, but:

A serious current issue is: that if wlan_assoc is to be set for a
device row, upon insertion, a FK-constraint is not satisfied since the
wlan processing is _after_ device processing and the wlan does not yet
exist.

Another issue is that the iframe in the 'Parser Log' option does not see
the changes made on disk and has to be manually reloaded (...or via
js!).