Execute multy-queries if not enpty
For malformed packets, the multy queries could end up having nothing to
update the DB with. So an empty query errs the connection.

Also a silly syntax error breaking the populate in the previous commit
was fixed.

For now it looks like the 'Populate' functionality is complete and
working. About to tag version before proceeding to fcaps section.