Finish Fault-Feature#1
* Completed the frontend-code,
* Implemented the backend code for DB access & calculations they call
* Styled the webpage returned

The last job introduced overriding stylesheets for pages demanding it.

Fixed some bugs in populate, bringing it to a level of completeness
where wifi.xml is succesfuly parsed, error-free, with expected results
reflected in the DB afterwards.

A final small touch, rules that SQL terms will be capitalized throughout
the queries to separate it from user-input (tables,columns,rows...).