Save final DB state
Prior to a local cleanup, a dump of all tables and data had to be
exported.

The .sql file had a few MB size, therefore compression was necessary to
leverage the SQL code redundancy and repetition.

A total of more than 23 thousands of packets, dozens of networks and
devices had been detected and extensive relevant info was stored in the
database. It would be a shame to lose all this data :)